tis_gap_app/requirements.txt
flask>=3.0.0
openai>=1.30.0
anthropic>=0.28.0
python-docx>=1.1.0
